+++ apps/web-vite/CHANGELOG.md
@@ -1,5 +1,26 @@
 # web-vite
 
+## 0.0.17
+
+### Patch Changes
+
+- Updated dependencies
+  - @hbwow/hooks@2.2.2
+
+## 0.0.16
+
+### Patch Changes
+
+- Updated dependencies
+  - @hbwow/hooks@2.2.1
+
+## 0.0.15
+
+### Patch Changes
+
+- Updated dependencies
+  - @hbwow/hooks@2.2.0
+
 ## 0.0.14
 
 ### Patch Changes